Flatten the request styles
